разделил DTO на версии для десктопа и мобильной и веб версии приложения
