+++ MedicineProject.Domain/DTOs/WebMobile/PatientDTO.cs
@@ -1,0 +1,7 @@
+﻿namespace MedicineProject.Domain.DTOs.WebMobile
+{
+    public record PatientDTO : UserDTO
+    {
+        public PatientDTO() { }
+    }
+}
