+++ MedicineProject.Domain/MedicineProject.Domain.csproj
@@ -18,4 +18,8 @@
     <PackageReference Include="Npgsql.EntityFrameworkCore.PostgreSQL" Version="7.0.4" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="DTOs\Desktop\" />
+  </ItemGroup>
+
 </Project>
