+++ MedicineProject.Domain/Models/WebMobile/Hospital.cs
@@ -1,4 +1,4 @@
-﻿using MedicineProject.Domain.DTOs;
+﻿using MedicineProject.Domain.DTOs.WebMobile;
 using MedicineProject.Domain.Models.Base;
 
 namespace MedicineProject.Domain.Models.WebMobile
